Email utility refactoring
