+++ .env.example
@@ -24,6 +24,7 @@ EMAIL_USE_TLS = False
 # Only add the "k" value from your ULS web token, not the entire JSON object!
 ULS_K_VALUE = "vatusa_uls_jwk_k_value"
 API_KEY = "vatusa_api_key"
+ULS_REDIR_URL = 6
 
 # Add all controlled (i.e., towered) fields in your ARTCC here.
 # If your CTR/FSS positions use a prefix that does not align with one of your airport codes
