+++ templates/partials/footer.html
@@ -12,9 +12,9 @@
                     <ul>
                         <li><a href="/">Home</a></li>
                         <li><a href="https://www.zidtraining.org/">Training</a></li>
-                        <li><a href="#">Leave Feedback</a></li>
-                        <li><a href="#">Events</a></li>
-                        <li><a href="#">IDS</a></li>
+                        <li><a href="/feedback/">Leave Feedback</a></li>
+                        <li><a href="/events/">Events</a></li>
+                        <!--<li><a href="#">IDS</a></li>-->
                     </ul>
                 </div>
